(converter) Move DomSampleClassifier to converter's code tree
